Move FilePtrSetStruct declaration into own header
And we don't install this header.  This is technically a backwards
incompatible change, but I suspect clients aren't using QUESO's
file-handling capabilities.

Now we don't like HDF5 internals in QUESO headers.
