+++ inc/queso/rebuild_makefile.sh
@@ -3,7 +3,7 @@
 # This file was shamelessly stolen from the libmesh source tree
 # See https://github.com/libMesh/libmesh
 
-headers=`find ../../src -name "*.h" -a -not -name all.h -a -not -name FiniteDistribution.h -a -not -name DiscreteSubset.h -type f | grep -v 'ANN' | grep -v 'interface' | LC_COLLATE=POSIX sort`
+headers=`find ../../src -name "*.h" -a -not -name all.h -a -not -name FiniteDistribution.h -a -not -name DiscreteSubset.h -a -not -name FilePtr.h -type f | grep -v 'ANN' | grep -v 'interface' | LC_COLLATE=POSIX sort`
 
 cat <<EOF > Makefile.am
 # Note - this file is automatically generated by $0
