+++ src/basic/src/SequenceOfVectors.C
@@ -25,6 +25,7 @@
 #include <queso/SequenceOfVectors.h>
 #include <queso/GslVector.h>
 #include <queso/GslMatrix.h>
+#include "core/inc/FilePtr.h"
 
 namespace QUESO {
