+++ src/stats/src/MonteCarloSG.C
@@ -25,6 +25,7 @@
 #include <queso/MonteCarloSG.h>
 #include <queso/GslVector.h>
 #include <queso/GslMatrix.h>
+#include "core/inc/FilePtr.h"
 
 namespace QUESO {
