+++ src/stats/src/StatisticalForwardProblem.C
@@ -26,6 +26,7 @@
 #include <queso/SequentialVectorRealizer.h>
 #include <queso/GslVector.h>
 #include <queso/GslMatrix.h>
+#include "core/inc/FilePtr.h"
 
 namespace QUESO {
